splite static files and infra code
branch: tianyu-shell-dev
user: victor.yao
