+++ ty-home/i18n/MessageBundle.ts
@@ -1,6 +1,6 @@
 /**@format */
 
-import { getTextFromFile } from "../../ty-common/shell/infra/message/Message";
+import { getTextFromFile } from "ty-infra/core/message/Message";
 
 export function getText(id: string, args?: (string | number)[] | string): string {
     return getTextFromFile("home", "i18n/message", id, args);
