+++ ty-infra/i18n/MessageBundle.ts
@@ -1,6 +1,6 @@
 /**@format */
 
-import { getTextFromFile } from "../infra/message/Message";
+import { getTextFromFile } from "ty-infra/core/message/Message";
 
 export function getText(id: string, args?: (string | number)[] | string): string {
     return getTextFromFile("infra", "i18n/message", id, args);
