+++ ty-infra/ui/css/waitingDialog.css
@@ -16,7 +16,7 @@
 }
 
 .wait_dialog_t {
-  color: var(--ts_ui_blk_4);
+  color: var(--ts_ui_blk_1);
   margin-left: auto;
   margin-right: auto;
